requirements.txt
beautifulsoup4
requests
python-dotenv
langchain
langchain-openai
scrapingbee